Moved event selction from entry level to figure
