--- src/views/Entry/EntryForm/EntryCloneForm/styles.css
@@ -1,30 +1,0 @@
-.form {
-    display: flex;
-    position: relative;
-    flex-direction: column;
-    overflow-y: auto;
-
-    .content {
-        display: flex;
-        flex-direction: column;
-        flex-grow: 1;
-        overflow-y: auto;
-
-        .new-event {
-            display: flex;
-            justify-content: flex-end;
-        }
-    }
-
-    .form-buttons {
-        display: flex;
-        flex-shrink: 0;
-        justify-content: flex-end;
-        margin-top: var(--spacing-medium);
-
-        .button {
-            margin-left: var(--spacing-medium);
-            padding: var(--spacing-small) var(--spacing-large);
-        }
-    }
-}
