+++ src/views/Entry/EntryForm/FigureInput/index.tsx
@@ -15,6 +15,7 @@ import {
     Switch,
     SelectInput,
     Button,
+    Modal,
 } from '@togglecorp/toggle-ui';
 import { isDefined, sum } from '@togglecorp/fujs';
 import {
@@ -28,6 +29,7 @@ import {
     gql,
     useQuery,
 } from '@apollo/client';
+import { IoMdEye, IoMdEyeOff } from 'react-icons/io';
 import { v4 as uuidv4 } from 'uuid';
 
 import MarkdownEditor from '#components/MarkdownEditor';
@@ -38,9 +40,12 @@ import NonFieldError from '#components/NonFieldError';
 import NonFieldWarning from '#components/NonFieldWarning';
 import Section from '#components/Section';
 import Header from '#components/Header';
+import useModalState from '#hooks/useModalState';
+import EventForm from '#components/forms/EventForm';
+import DomainContext from '#components/DomainContext';
 import TrafficLightInput from '#components/TrafficLightInput';
-import { CountryOption } from '#components/selections/CountrySelectInput';
 import FigureTagMultiSelectInput, { FigureTagOption } from '#components/selections/FigureTagMultiSelectInput';
+import EventSelectInput, { EventOption } from '#components/selections/EventSelectInput';
 
 import {
     enumKeySelector,
@@ -119,9 +124,9 @@ interface FigureInputProps {
     onReviewChange?: (newValue: EntryReviewStatus, name: string) => void;
     trafficLightShown: boolean;
 
-    countries: CountryOption[] | null | undefined;
     selected?: boolean;
-
+    events: EventOption[] | null | undefined;
+    setEvents: Dispatch<SetStateAction<EventOption[] | null | undefined>>;
     tagOptions: TagOptions;
     setTagOptions: Dispatch<SetStateAction<FigureTagOption[] | null | undefined>>;
     optionsDisabled: boolean;
@@ -148,14 +153,15 @@ function FigureInput(props: FigureInputProps) {
         disabled,
         mode,
         review,
+        events,
         onReviewChange,
 
-        countries,
         selected,
 
         optionsDisabled: figureOptionsDisabled,
         tagOptions,
         setTagOptions,
+        setEvents,
         accuracyOptions,
         identifierOptions,
         categoryOptions,
@@ -171,17 +177,28 @@ function FigureInput(props: FigureInputProps) {
     } = props;
 
     const { notify } = useContext(NotificationContext);
+    const { user } = useContext(DomainContext);
+    const eventPermissions = user?.permissions?.event;
 
     const [selectedAge, setSelectedAge] = useState<string | undefined>();
     const [mapShown, setMapShown] = useState<boolean | undefined>(false);
+    const [eventDetailsShown, , , , toggleEventDetailsShown] = useModalState(false);
+
+    const [
+        shouldShowEventModal,
+        eventModalId,
+        showEventModal,
+        hideEventModal,
+    ] = useModalState();
 
     const editMode = mode === 'edit';
     const reviewMode = mode === 'review';
+    const eventNotChosen = !value.event;
 
     const { country, startDate } = value;
-    const year = startDate?.match(/^\d+/)?.[0];
+    const year = (() => (startDate?.match(/^\d+/)?.[0]))();
 
-    const variables = React.useMemo(
+    const variables = useMemo(
         () => (
             year && country
                 ? {
@@ -205,6 +222,19 @@ function FigureInput(props: FigureInputProps) {
     const households = [householdData?.householdSize].filter(isDefined);
 
     const onValueChange = useFormObject(index, onChange, defaultValue);
+
+    const handleEventCreate = useCallback(
+        (newEvent: EventOption) => {
+            setEvents((oldEvents) => [...(oldEvents ?? []), newEvent]);
+            onValueChange(newEvent.id, 'event' as const);
+            hideEventModal();
+        },
+        [
+            onValueChange,
+            hideEventModal,
+            setEvents,
+        ],
+    );
 
     const handleCountryChange = useCallback(
         (countryValue: string | undefined, countryName: 'country') => {
@@ -258,7 +288,9 @@ function FigureInput(props: FigureInputProps) {
     // FIXME: The type of value should have be FigureInputValueWithId instead.
     const { id: figureId } = value as FigureInputValueWithId;
 
-    const currentCountry = countries?.find((item) => item.id === value.country);
+    const selectedEvent = events?.find((item) => item.id === value?.event);
+    // FIXME: The value "countries" of selectedEvent needs to be handled from server.
+    const currentCountry = selectedEvent?.countries.find((item) => item.id === value?.country);
     const currentCategory = value.category as (
         FigureCategoryTypes | undefined);
     const currentTerm = value.term as (
@@ -296,6 +328,77 @@ function FigureInput(props: FigureInputProps) {
         ? totalValue - totalDisaggregatedValue
         : 0;
 
+    const eventBlock = (
+        <>
+            <Row>
+                <div className={styles.eventRow}>
+                    <EventSelectInput
+                        error={error?.fields?.event}
+                        label="Event *"
+                        name="event"
+                        className={styles.eventSelectInput}
+                        options={events}
+                        value={value.event}
+                        onChange={onValueChange}
+                        onOptionsChange={setEvents}
+                        disabled={disabled || figureOptionsDisabled}
+                        readOnly={!editMode}
+                        icons={trafficLightShown && review && (
+                            <TrafficLightInput
+                                disabled={!reviewMode}
+                                name="event"
+                                onChange={onReviewChange}
+                                value={review.event?.value}
+                                comment={review.event?.comment}
+                            />
+                        )}
+                        actions={(
+                            <Button
+                                onClick={toggleEventDetailsShown}
+                                name={undefined}
+                                transparent
+                                title={eventDetailsShown ? 'Hide event details' : 'Show event details'}
+                                compact
+                            >
+                                {eventDetailsShown ? <IoMdEyeOff /> : <IoMdEye />}
+                            </Button>
+                        )}
+                    />
+                    {eventPermissions && (
+                        <Button
+                            name={undefined}
+                            className={styles.addEventButton}
+                            onClick={showEventModal}
+                        >
+                            Add Event
+                        </Button>
+                    )}
+                </div>
+            </Row>
+            {shouldShowEventModal && (
+                <Modal
+                    className={styles.addEventModal}
+                    bodyClassName={styles.body}
+                    heading="Add Event"
+                    onClose={hideEventModal}
+                >
+                    <EventForm
+                        id={eventModalId}
+                        onEventCreate={handleEventCreate}
+                        onEventFormCancel={hideEventModal}
+                    />
+                </Modal>
+            )}
+            {value.event && eventDetailsShown && (
+                <EventForm
+                    className={styles.eventDetails}
+                    id={value.event}
+                    readOnly
+                />
+            )}
+        </>
+    );
+
     return (
         <Section
             elementRef={elementRef}
@@ -317,17 +420,18 @@ function FigureInput(props: FigureInputProps) {
             <NonFieldError>
                 {error?.$internal}
             </NonFieldError>
+            {eventBlock}
             <Row>
                 <SelectInput
                     error={error?.fields?.country}
                     label="Country *"
                     name="country"
-                    options={countries}
+                    options={selectedEvent?.countries}
                     value={value.country}
                     keySelector={countryKeySelector}
                     labelSelector={countryLabelSelector}
                     onChange={handleCountryChange}
-                    disabled={disabled}
+                    disabled={disabled || eventNotChosen}
                     // NOTE: Disable changing country when there are more than one geolocation
                     readOnly={!editMode || (value.geoLocations?.length ?? 0) > 0}
                     nonClearable
@@ -342,6 +446,7 @@ function FigureInput(props: FigureInputProps) {
                         <Button
                             name={undefined}
                             onClick={handleShowMapAction}
+                            disabled={eventNotChosen}
                             compact
                             transparent
                         >
@@ -359,7 +464,7 @@ function FigureInput(props: FigureInputProps) {
                         onChange={onValueChange}
                         country={currentCountry}
                         readOnly={!editMode}
-                        disabled={disabled}
+                        disabled={disabled || eventNotChosen}
                     />
                 </Row>
             )}
@@ -376,7 +481,7 @@ function FigureInput(props: FigureInputProps) {
                             onChange={onGeoLocationChange}
                             onRemove={onGeoLocationRemove}
                             error={error?.fields?.geoLocations?.members?.[geoLocation.uuid]}
-                            disabled={disabled}
+                            disabled={disabled || eventNotChosen}
                             mode={mode}
                             review={review}
                             onReviewChange={onReviewChange}
@@ -395,7 +500,7 @@ function FigureInput(props: FigureInputProps) {
                     onChange={onValueChange}
                     value={value.calculationLogic}
                     error={error?.fields?.calculationLogic}
-                    disabled={disabled}
+                    disabled={disabled || eventNotChosen}
                     readOnly={!editMode}
                     icons={trafficLightShown && review && (
                         <TrafficLightInput
@@ -413,7 +518,7 @@ function FigureInput(props: FigureInputProps) {
                     onChange={onValueChange}
                     value={value.caveats}
                     error={error?.fields?.caveats}
-                    disabled={disabled}
+                    disabled={disabled || eventNotChosen}
                     readOnly={!editMode}
                     icons={trafficLightShown && review && (
                         <TrafficLightInput
@@ -431,7 +536,7 @@ function FigureInput(props: FigureInputProps) {
                     value={value.sourceExcerpt}
                     name="sourceExcerpt"
                     error={error?.fields?.sourceExcerpt}
-                    disabled={disabled}
+                    disabled={disabled || eventNotChosen}
                     readOnly={!editMode}
                     icons={trafficLightShown && review && (
                         <TrafficLightInput
@@ -450,7 +555,7 @@ function FigureInput(props: FigureInputProps) {
                     onChange={onValueChange}
                     value={value.tags}
                     error={error?.fields?.tags?.$internal}
-                    disabled={disabled || figureOptionsDisabled}
+                    disabled={disabled || figureOptionsDisabled || eventNotChosen}
                     readOnly={!editMode}
                     onOptionsChange={setTagOptions}
                 />
@@ -465,7 +570,7 @@ function FigureInput(props: FigureInputProps) {
                     value={value.category}
                     onChange={onValueChange}
                     error={error?.fields?.category}
-                    disabled={disabled || figureOptionsDisabled}
+                    disabled={disabled || figureOptionsDisabled || eventNotChosen}
                     readOnly={!editMode}
                     icons={trafficLightShown && review && (
                         <TrafficLightInput
@@ -484,7 +589,7 @@ function FigureInput(props: FigureInputProps) {
                     value={value.role}
                     onChange={onValueChange}
                     error={error?.fields?.role}
-                    disabled={disabled || figureOptionsDisabled}
+                    disabled={disabled || figureOptionsDisabled || eventNotChosen}
                     readOnly={!editMode}
                     icons={trafficLightShown && review && (
                         <TrafficLightInput
@@ -503,7 +608,7 @@ function FigureInput(props: FigureInputProps) {
                     value={value.term}
                     onChange={onValueChange}
                     error={error?.fields?.term}
-                    disabled={disabled || figureOptionsDisabled}
+                    disabled={disabled || figureOptionsDisabled || eventNotChosen}
                     readOnly={!editMode}
                     icons={trafficLightShown && review && (
                         <TrafficLightInput
@@ -523,6 +628,7 @@ function FigureInput(props: FigureInputProps) {
                         value={value.displacementOccurred}
                         onChange={onValueChange}
                         error={error?.fields?.displacementOccurred}
+                        disabled={eventNotChosen}
                         readOnly={!editMode}
                         icons={trafficLightShown && review && (
                             <TrafficLightInput
@@ -544,7 +650,7 @@ function FigureInput(props: FigureInputProps) {
                     value={value.quantifier}
                     onChange={onValueChange}
                     error={error?.fields?.quantifier}
-                    disabled={disabled || figureOptionsDisabled}
+                    disabled={disabled || figureOptionsDisabled || eventNotChosen}
                     readOnly={!editMode}
                     icons={trafficLightShown && review && (
                         <TrafficLightInput
@@ -560,7 +666,7 @@ function FigureInput(props: FigureInputProps) {
                     value={value.reported}
                     onChange={onValueChange}
                     error={error?.fields?.reported}
-                    disabled={disabled}
+                    disabled={disabled || eventNotChosen}
                     readOnly={!editMode}
                     icons={trafficLightShown && review && (
                         <TrafficLightInput
@@ -579,7 +685,7 @@ function FigureInput(props: FigureInputProps) {
                     value={value.unit}
                     onChange={onValueChange}
                     error={error?.fields?.unit}
-                    disabled={disabled || figureOptionsDisabled}
+                    disabled={disabled || figureOptionsDisabled || eventNotChosen}
                     readOnly={!editMode}
                     icons={trafficLightShown && review && (
                         <TrafficLightInput
@@ -597,7 +703,7 @@ function FigureInput(props: FigureInputProps) {
                             value={value.householdSize}
                             onChange={onValueChange}
                             error={error?.fields?.householdSize}
-                            disabled={disabled}
+                            disabled={disabled || eventNotChosen}
                             readOnly={!editMode}
                             suggestions={households}
                             suggestionKeySelector={householdKeySelector}
@@ -607,7 +713,7 @@ function FigureInput(props: FigureInputProps) {
                             label="Total Figure"
                             name="totalFigure"
                             value={Math.floor((value.householdSize ?? 0) * (value.reported ?? 0))}
-                            disabled={disabled}
+                            disabled={disabled || eventNotChosen}
                             readOnly
                         />
                     </>
@@ -619,7 +725,7 @@ function FigureInput(props: FigureInputProps) {
                     name="startDate"
                     value={value.startDate}
                     onChange={onValueChange}
-                    disabled={disabled}
+                    disabled={disabled || eventNotChosen}
                     error={error?.fields?.startDate}
                     readOnly={!editMode}
                     icons={trafficLightShown && review && (
@@ -639,7 +745,7 @@ function FigureInput(props: FigureInputProps) {
                     value={value.startDateAccuracy}
                     onChange={onValueChange}
                     error={error?.fields?.startDateAccuracy}
-                    disabled={disabled || figureOptionsDisabled}
+                    disabled={disabled || figureOptionsDisabled || eventNotChosen}
                     readOnly={!editMode}
                     icons={trafficLightShown && review && (
                         <TrafficLightInput
@@ -654,7 +760,7 @@ function FigureInput(props: FigureInputProps) {
                     name="endDate"
                     value={value.endDate}
                     onChange={onValueChange}
-                    disabled={disabled}
+                    disabled={disabled || eventNotChosen}
                     error={error?.fields?.endDate}
                     readOnly={!editMode}
                     icons={trafficLightShown && review && (
@@ -675,7 +781,7 @@ function FigureInput(props: FigureInputProps) {
                         value={value.endDateAccuracy}
                         onChange={onValueChange}
                         error={error?.fields?.endDateAccuracy}
-                        disabled={disabled || figureOptionsDisabled}
+                        disabled={disabled || figureOptionsDisabled || eventNotChosen}
                         readOnly={!editMode}
                         icons={trafficLightShown && review && (
                             <TrafficLightInput
@@ -704,7 +810,7 @@ function FigureInput(props: FigureInputProps) {
                         value={value.isHousingDestruction}
                         onChange={onValueChange}
                         // error={error?.fields?.isHousingDestruction}
-                        disabled={disabled}
+                        disabled={disabled || eventNotChosen}
                         readOnly={!editMode}
                     />
                 </Row>
@@ -725,7 +831,7 @@ function FigureInput(props: FigureInputProps) {
                     value={value.isDisaggregated}
                     onChange={onValueChange}
                     // error={error?.fields?.isDisaggregated}
-                    disabled={disabled}
+                    disabled={disabled || eventNotChosen}
                     readOnly={!editMode}
                 />
             </Row>
@@ -738,7 +844,7 @@ function FigureInput(props: FigureInputProps) {
                             value={value.disaggregationDisplacementUrban}
                             onChange={onValueChange}
                             error={error?.fields?.disaggregationDisplacementUrban}
-                            disabled={disabled}
+                            disabled={disabled || eventNotChosen}
                             readOnly={!editMode}
                             icons={trafficLightShown && review && (
                                 <TrafficLightInput
@@ -754,7 +860,7 @@ function FigureInput(props: FigureInputProps) {
                             value={value.disaggregationDisplacementRural}
                             onChange={onValueChange}
                             error={error?.fields?.disaggregationDisplacementRural}
-                            disabled={disabled}
+                            disabled={disabled || eventNotChosen}
                             readOnly={!editMode}
                             icons={trafficLightShown && review && (
                                 <TrafficLightInput
@@ -772,7 +878,7 @@ function FigureInput(props: FigureInputProps) {
                             value={value.disaggregationLocationCamp}
                             onChange={onValueChange}
                             error={error?.fields?.disaggregationLocationCamp}
-                            disabled={disabled}
+                            disabled={disabled || eventNotChosen}
                             readOnly={!editMode}
                             icons={trafficLightShown && review && (
                                 <TrafficLightInput
@@ -788,7 +894,7 @@ function FigureInput(props: FigureInputProps) {
                             value={value.disaggregationLocationNonCamp}
                             onChange={onValueChange}
                             error={error?.fields?.disaggregationLocationNonCamp}
-                            disabled={disabled}
+                            disabled={disabled || eventNotChosen}
                             readOnly={!editMode}
                             icons={trafficLightShown && review && (
                                 <TrafficLightInput
@@ -806,7 +912,7 @@ function FigureInput(props: FigureInputProps) {
                             value={value.disaggregationDisability}
                             onChange={onValueChange}
                             error={error?.fields?.disaggregationDisability}
-                            disabled={disabled}
+                            disabled={disabled || eventNotChosen}
                             readOnly={!editMode}
                             icons={trafficLightShown && review && (
                                 <TrafficLightInput
@@ -822,7 +928,7 @@ function FigureInput(props: FigureInputProps) {
                             value={value.disaggregationIndigenousPeople}
                             onChange={onValueChange}
                             error={error?.fields?.disaggregationIndigenousPeople}
-                            disabled={disabled}
+                            disabled={disabled || eventNotChosen}
                             readOnly={!editMode}
                             icons={trafficLightShown && review && (
                                 <TrafficLightInput
@@ -841,7 +947,7 @@ function FigureInput(props: FigureInputProps) {
                                 <Button
                                     name={undefined}
                                     onClick={handleAgeAdd}
-                                    disabled={disabled}
+                                    disabled={disabled || eventNotChosen}
                                 >
                                     Add Age & Gender
                                 </Button>
@@ -877,7 +983,7 @@ function FigureInput(props: FigureInputProps) {
                                 error={
                                     error?.fields?.disaggregationAge?.members?.[age.uuid]
                                 }
-                                disabled={disabled}
+                                disabled={disabled || eventNotChosen}
                                 mode={mode}
                                 review={review}
                                 onReviewChange={onReviewChange}
@@ -894,7 +1000,7 @@ function FigureInput(props: FigureInputProps) {
                     name="includeIdu"
                     value={value.includeIdu}
                     onChange={onValueChange}
-                    disabled={disabled}
+                    disabled={disabled || eventNotChosen}
                     readOnly={!editMode}
                 />
             </Row>
@@ -905,7 +1011,7 @@ function FigureInput(props: FigureInputProps) {
                         name="excerptIdu"
                         value={value.excerptIdu}
                         onChange={onValueChange}
-                        disabled={disabled}
+                        disabled={disabled || eventNotChosen}
                         error={error?.fields?.excerptIdu}
                         readOnly={!editMode}
                         icons={trafficLightShown && review && (
