+++ src/views/Event/index.tsx
@@ -27,7 +27,6 @@ import {
     EventSummaryQueryVariables,
 } from '#generated/types';
 
-import EventCloneForm from './EventCloneForm';
 import styles from './styles.css';
 
 const EVENT = gql`
@@ -107,13 +106,6 @@ function Event(props: EventProps) {
     const eventPermissions = user?.permissions?.event;
 
     const [
-        shouldShowEventCloneModal,
-        eventIdToClone,
-        showEventCloneModal,
-        hideEventCloneModal,
-    ] = useModalState();
-
-    const [
         shouldShowAddEventModal,
         editableEventId,
         showAddEventModal,
@@ -122,8 +114,7 @@ function Event(props: EventProps) {
 
     const [
         alertShown,
-        clonedEvent,
-        showAlert,
+        clonedEvent,,
         hideAlert,
     ] = useModalState<string>(false);
 
@@ -133,23 +124,6 @@ function Event(props: EventProps) {
         const { name } = eventData.event;
         title = crisisName ? `${crisisName} › ${name}` : name;
     }
-
-    const handleCloneEntryButtonClick = React.useCallback(
-        () => {
-            showEventCloneModal(eventId);
-        },
-        [eventId, showEventCloneModal],
-    );
-
-    const handleCloneModalClose = React.useCallback(
-        (event?: string) => {
-            if (event) {
-                showAlert(event);
-            }
-            hideEventCloneModal();
-        },
-        [showAlert, hideEventCloneModal],
-    );
 
     const handleAlertAction = React.useCallback(
         () => {
@@ -173,28 +147,14 @@ function Event(props: EventProps) {
                 className={styles.container}
                 contentClassName={styles.details}
                 heading="Details"
-                headerActions={(
-                    <>
-                        {eventPermissions?.add && (
-                            <Button
-                                name={undefined}
-                                variant="default"
-                                onClick={handleCloneEntryButtonClick}
-                                disabled={loading || !eventData?.event?.id}
-                            >
-                                Clone
-                            </Button>
-                        )}
-                        {eventPermissions?.change && (
-                            <Button
-                                name={eventData?.event?.id}
-                                onClick={showAddEventModal}
-                                disabled={loading || !eventData?.event?.id}
-                            >
-                                Edit Event
-                            </Button>
-                        )}
-                    </>
+                headerActions={eventPermissions?.change && (
+                    <Button
+                        name={eventData?.event?.id}
+                        onClick={showAddEventModal}
+                        disabled={loading || !eventData?.event?.id}
+                    >
+                        Edit Event
+                    </Button>
                 )}
             >
                 {eventData ? (
@@ -282,7 +242,6 @@ function Event(props: EventProps) {
             </Container>
             <EntriesTable
                 className={styles.largeContainer}
-                eventId={eventId}
                 eventColumnHidden
                 crisisColumnHidden
             />
@@ -295,17 +254,6 @@ function Event(props: EventProps) {
                         id={editableEventId}
                         onEventCreate={hideAddEventModal}
                         onEventFormCancel={hideAddEventModal}
-                    />
-                </Modal>
-            )}
-            {shouldShowEventCloneModal && eventId && (
-                <Modal
-                    onClose={hideEventCloneModal}
-                    heading="Clone Event"
-                >
-                    <EventCloneForm
-                        eventId={eventIdToClone}
-                        onCloseForm={handleCloneModalClose}
                     />
                 </Modal>
             )}
