+++ src/views/Extraction/ExtractionEntriesTable/NudeFigureTable/index.tsx
@@ -255,7 +255,7 @@ function NudeFigureTable(props: NudeFigureTableProps) {
                     'entry__event__crisis__name',
                     'Crisis',
                     (item) => ({
-                        title: item.event.crisis?.name,
+                        title: item.event?.crisis?.name,
                         attrs: { crisisId: item.event?.crisis?.id },
                         ext: undefined,
                     }),
@@ -295,7 +295,7 @@ function NudeFigureTable(props: NudeFigureTableProps) {
                 createTextColumn<FigureFields, string>(
                     'entry__event__event_type',
                     'Cause',
-                    (item) => item.event.eventType,
+                    (item) => item.event?.eventType,
                     { sortable: true },
                 ),
                 createTextColumn<FigureFields, string>(
@@ -321,7 +321,7 @@ function NudeFigureTable(props: NudeFigureTableProps) {
                     'Figure Category',
                     (item) => ({
                         title: item.category,
-                        attrs: { eventId: item.event.id },
+                        attrs: { eventId: item.event?.id },
                         ext: item.oldId
                             ? `/facts/${item.oldId}`
                             : undefined,
