+++ src/views/Report/ReportEntryTable/index.tsx
@@ -11,7 +11,6 @@ import {
 } from '@togglecorp/toggle-ui';
 import {
     createTextColumn,
-    createLinkColumn,
     createStatusColumn,
     createDateColumn,
     createNumberColumn,
@@ -48,7 +47,7 @@ const GET_REPORT_ENTRIES_LIST = gql`
             ordering: $ordering,
             page: $page,
             pageSize: $pageSize,
-            ) {
+        ) {
                 totalCount
                 results {
                     totalFlowNdFigures
@@ -206,31 +205,35 @@ function ReportEntryTable(props: ReportEntryProps) {
                 (item) => item.createdBy?.fullName,
                 { sortable: true },
             ),
-            createLinkColumn<ReportEntryFields, string>(
-                'event__crisis__name',
-                'Crisis',
-                (item) => ({
-                    title: item.figures.event?.crisis?.name,
-                    attrs: { crisisId: item.figures.event?.crisis?.id },
-                    ext: undefined,
-                }),
-                route.crisis,
-                { sortable: true },
-            ),
-            createLinkColumn<ReportEntryFields, string>(
-                'event__name',
-                'Event',
-                (item) => ({
-                    title: item.figures.event?.name,
-                    // FIXME: this may be wrong
-                    attrs: { eventId: item.figures.event?.id },
-                    ext: item.figures.event?.oldId
-                        ? `/events/${item.figures.event.oldId}`
-                        : undefined,
-                }),
-                route.event,
-                { sortable: true },
-            ),
+            /* FIX: Below the field "figures" in each item comes as an array which needs to be fixed
+             in order to display as a short data instead of joining
+             all the elements of that array */
+
+            /* createLinkColumn<ReportEntryFields, string>(
+                 'event__crisis__name',
+                 'Crisis',
+                 (item) => ({
+                     title: item.figures.event?.crisis?.name,
+                     attrs: { crisisId: item.figures.event?.crisis?.id },
+                     ext: undefined,
+                 }),
+                 route.crisis,
+                 { sortable: true },
+             ),
+             createLinkColumn<ReportEntryFields, string>(
+                 'event__name',
+                 'Event',
+                 (item) => ({
+                     title: item.figures.event?.name,
+                     // FIXME: this may be wrong
+                     attrs: { eventId: item.figures.event?.id },
+                     ext: item.figures.event?.oldId
+                         ? `/events/${item.figures.event.oldId}`
+                         : undefined,
+                 }),
+                 route.event,
+                 { sortable: true },
+             ), */
             createStatusColumn<ReportEntryFields, string>(
                 'article_title',
                 'Entry',
